refactor: service 로직 하나로 합침.
